kubernetes.rb: add restrict_to_default_namespace flag
At the moment, the plugin scans the entire cluster for pods and
jobs that have been created. This means the service the plugin is
running on **must** have 'cluster-wide' permissions.

This commit changes that by scoping the list operations for pods
and jobs to the 'default' namespace the plugin is configured against.